TicketDAOTest readyById + readAll done Changed ApiException

+++ src/main/java/dat/dtos/MessageDTO.java
@@ -13,7 +13,7 @@ import java.util.stream.Collectors;
 @NoArgsConstructor
 public class MessageDTO {
     private Long id;
-    private Long ticketId;
+    private int ticketId;
     private UserDTO author;
     private boolean internalFlag;
     private String body;
